# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "System Friction Institute External Agent API",
    "version": "1.8.0",
    "description": "Governed machine interface for authorized AI clients. GPT Actions authenticate with SFI OAuth. Institutional principals receive explicitly configured authority; normal accounts receive only owner-scoped personal Lab, Cognitive and Studio capabilities.",
    "termsOfService": "https://systemfriction.org/privacy"
  },
...
  "servers": [
    {
      "url": "https://systemfriction.org"
    }
  ],
  "security": [
    {
      "sfiOAuth": []
    }
  ],
  "paths": {
...
    "/api/external/v1/personal-lab": {
      "post": {
        "operationId": "operatePersonalLabWorkspace",
        "summary": "Read, create and run owner-scoped personal Lab work",
        "description": "OAuth subject-bound. state/report require lab:read; create_case/persist require lab:write; run requires lab:run.",
        "security": [
          {
            "sfiOAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalLabRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Personal Lab state or report"
          },
          "201": {
            "description": "Owned case/evidence/run persisted"
          },
          "207": {
            "description": "Run completed with degraded automation results"
          },
          "400": {
            "description": "Invalid operation"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "OAuth subject required"
          },
          "404": {
            "description": "Owned case/evidence not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PersonalLabRequest": {
        "type": "object",
        "required": [
          "operation"
        ],
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "state",
              "report",
              "create_case",
              "persist",
              "run"
            ]
          },
          "title": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "baseline": {
            "type": "string"
          },
          "verificationWindow": {
            "type": "string",
            "enum": [
              "72h",
              "7d",
              "30d"
            ]
          },
          "caseId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "reliability": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "evidenceType": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          },
          "observedAt": {
            "type": "string"
          },
          "protocolId": {
            "type": "string",
            "enum": [
              "sociotechnical_simulation",
              "economic_simulation"
            ]
          },
          "evidenceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "parameters": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sfiOAuth": {
        "type": "oauth2",
        "description": "User-bound SFI identity for GPT Actions. Institutional accounts may receive explicit institutional scopes; normal accounts are reduced to owner-scoped personal workspace capabilities.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://systemfriction.org/api/oauth/authorize",
            "tokenUrl": "https://systemfriction.org/api/oauth/token",
            "scopes": {
              "observe": "Read governed institutional SFI state when the authenticated principal is allowed.",
              "propose": "Submit governed institutional proposals and evidence candidates for ROOT review.",
              "execute": "Execute only already-authorized queued institutional proposals and record proposal-scoped returns.",
              "lab:read": "Read Method Lab or personal Cognitive/Lab state permitted to the OAuth tenant.",
              "lab:write": "Persist permitted structured or owner-scoped personal Lab records.",
              "lab:run": "Run supported Method Lab or owner-scoped personal Cognitive/Lab runtimes.",
              "studio:read": "List, inspect and read features for Studio objects owned by the OAuth principal.",
              "studio:content": "Obtain a short-lived signed URL for Studio content owned by the OAuth principal.",
              "studio:run": "Run the server-side Studio analyzer for owned audio/video objects."
            }
          }
        }
      }
    }
  },
...
}